refactor: add tiktoken as core dependency and minor refactoring
