+++ haystack/nodes/retriever/_base_embedding_encoder.py
