+++ pyproject.toml
@@ -87,6 +87,9 @@ dependencies = [
   # Elasticsearch
   "elasticsearch>=7.7,<8",
 
+  # OpenAI tokenizer
+  "tiktoken>=0.1.2; python_version >= '3.8'",
+
   # context matching
   "rapidfuzz>=2.0.15,<2.8.0",   # FIXME https://github.com/deepset-ai/haystack/pull/3199
 
@@ -173,9 +176,6 @@ onnx = [
 onnx-gpu = [
   "onnxruntime-gpu",
   "onnxruntime_tools",
-]
-openai = [
-  "tiktoken>=0.1.2; python_version >= '3.8'"
 ]
 ray = [
   "ray>=1.9.1,<2; platform_system != 'Windows'",
